+++ tests/Unit/RabbitDriverTest.php
@@ -18,7 +18,7 @@ class RabbitDriverTest extends TestCase
 
     private $queue = 'some.queue';
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();